use gradle's test suite feature to separate integration tests
