+++ org.lflang/core/src/integrationTest/java/org/lflang/tests/runtime/CSchedulerTest.java
@@ -22,7 +22,6 @@ public class CSchedulerTest extends TestBase {
    * (e.g., -Dscheduler=GEDF_NP).
    */
   @Test
-  @Tag("Integration")
   public void runWithNonDefaultSchedulers() {
     EnumSet<TestCategory> categories = EnumSet.of(TestCategory.CONCURRENT, TestCategory.MULTIPORT);
