+++ org.lflang/core/src/integrationTest/java/org/lflang/tests/runtime/CTest.java
@@ -24,7 +24,6 @@
  ***************/
 package org.lflang.tests.runtime;
 
-import java.util.Properties;
 import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
@@ -33,33 +32,18 @@ import org.lflang.Target;
 import org.lflang.tests.RuntimeTest;
 
 /**
- * Collection of tests for the Python target.
+ * Collection of tests for the C target.
  *
- * <p>Even though all tests are implemented in the base class, we override them here so that each
- * test can be easily invoked individually from IDEs with JUnit support like Eclipse and IntelliJ.
- * This is typically done by right-clicking on the name of the test method and then clicking "Run".
+ * <p>Tests that are implemented in the base class are still overridden so that each test can be
+ * easily invoked individually from IDEs with JUnit support like Eclipse and IntelliJ. This is
+ * typically done by right-clicking on the name of the test method and then clicking "Run".*
  *
  * @author Marten Lohstroh
  */
-public class PythonTest extends RuntimeTest {
+public class CTest extends RuntimeTest {
 
-  public PythonTest() {
-    super(Target.Python);
-  }
-
-  @Override
-  protected void addExtraLfcArgs(Properties args) {
-    super.addExtraLfcArgs(args);
-    if (System.getProperty("os.name").startsWith("Windows")) {
-      // Use the RelWithDebInfo build type on Windows as the Debug/Test build type produces linker
-      // Errors in CI
-      args.setProperty("build-type", "RelWithDebInfo");
-    }
-  }
-
-  @Override
-  protected boolean supportsFederatedExecution() {
-    return true;
+  public CTest() {
+    super(Target.C);
   }
 
   @Override
@@ -68,34 +52,42 @@ public class PythonTest extends RuntimeTest {
   }
 
   @Override
+  protected boolean supportsFederatedExecution() {
+    return true;
+  }
+
+  @Override
   protected boolean supportsDockerOption() {
-    return false; // FIXME: https://issues.lf-lang.org/1564
+    return true;
   }
 
   @Test
-  @Tag("Integration")
   @Override
   public void runGenericTests() {
     super.runGenericTests();
   }
 
   @Test
-  @Tag("Integration")
   @Override
   public void runTargetSpecificTests() {
+    Assumptions.assumeFalse(isWindows(), Message.NO_WINDOWS_SUPPORT);
     super.runTargetSpecificTests();
   }
 
   @Test
-  @Tag("Integration")
   @Override
   public void runMultiportTests() {
     super.runMultiportTests();
   }
 
   @Test
-  @Tag("Integration")
-  @Disabled("TODO")
+  @Override
+  public void runWithThreadingOff() {
+    super.runWithThreadingOff();
+  }
+
+  @Test
+  @Disabled("TODO only 27/96 tests pass")
   @Override
   public void runAsFederated() {
     Assumptions.assumeFalse(isWindows(), Message.NO_WINDOWS_SUPPORT);
@@ -103,14 +95,12 @@ public class PythonTest extends RuntimeTest {
   }
 
   @Test
-  @Tag("Integration")
   @Override
   public void runConcurrentTests() {
     super.runConcurrentTests();
   }
 
   @Test
-  @Tag("Integration")
   @Override
   public void runFederatedTests() {
     Assumptions.assumeFalse(isWindows(), Message.NO_WINDOWS_SUPPORT);
@@ -118,14 +108,22 @@ public class PythonTest extends RuntimeTest {
   }
 
   @Test
-  @Tag("Integration")
+  public void runModalTests() {
+    super.runModalTests();
+  }
+
+  @Test
+  public void runNoInliningTests() {
+    super.runNoInliningTests();
+  }
+
+  @Test
   @Override
   public void runDockerTests() {
     super.runDockerTests();
   }
 
   @Test
-  @Tag("Integration")
   @Override
   public void runDockerFederatedTests() {
     Assumptions.assumeFalse(isWindows(), Message.NO_WINDOWS_SUPPORT);
